Featurerequirements.txt
transformers==4.29.2 
torch==2.4.0 
biopython==1.83 
einops==0.8.0
gensim==3.8.3  
scipy==1.7.3 
pandas 
scikit-learn
Logbook==1.0.0  